Use Float for Stripe tax_percent
Stripe uses a Float in their examples, and that's what we used to use
(465a47469). Integer conversion introduced in (d655308).

> A non-negative decimal (with at most four decimal places) between 0
> and 100. This represents the percentage of the subscription invoice
> subtotal that will be calculated and added as tax to the final amount
> in each billing period. For example, a plan which charges $10/month
> with a tax_percent of 20.0 will charge $12 per invoice.
>
> https://stripe.com/docs/api/subscriptions/create#create_subscription-tax_percent

+++ app/controllers/alaveteli_pro/subscriptions_controller.rb
@@ -222,7 +222,7 @@ class AlaveteliPro::SubscriptionsController < AlaveteliPro::BaseController
   end
 
   def tax_percent
-    (BigDecimal(AlaveteliConfiguration.stripe_tax_rate).to_f * 100).to_i
+    (BigDecimal(AlaveteliConfiguration.stripe_tax_rate).to_f * 100).to_f
   end
 
   def check_plan_exists
